Move private code of HMC804x into its folder
